Sista library divided in .hpp headers and .cpp implementations
