+++ include/sista/border.hpp
@@ -10,13 +10,10 @@ namespace sista {
         ANSI::Settings settings; // settings of the Border style
 
     public:
-        Border(char symbol_, ANSI::Settings settings_): symbol(symbol_), settings(settings_) {}
-        Border(char symbol_, ANSI::Settings& settings_, bool _by_reference): symbol(symbol_), settings(settings_) {}
+        Border(char, ANSI::Settings);
+        Border(char, ANSI::Settings&, bool);
 
-        virtual void print(bool apply_settings=true) { // Print the Border
-            if (apply_settings)
-                settings.apply(); // Apply the settings
-            std::cout << symbol; // Print the symbol
-        }
+        virtual void print();
+        virtual void print(bool);
     };
 };
