+++ include/sista/coordinates.hpp
@@ -10,18 +10,12 @@ namespace sista {
         unsigned short y; // y coordinate
         unsigned short x; // x coordinate
 
-        Coordinates(): y(0), x(0) {} // Constructor
-        Coordinates(unsigned short y_, unsigned short x_): y(y_), x(x_) {} // Constructor
-        Coordinates(Coord coord): y(coord.first), x(coord.second) {} // Constructor
+        Coordinates();
+        Coordinates(unsigned short, unsigned short);
+        Coordinates(Coord);
 
-        bool operator==(const Coordinates& other) const {
-            return (y == other.y && x == other.x);
-        }
-        bool operator!=(const Coordinates& other) const {
-            return (y != other.y || x != other.x);
-        }
-        Coordinates operator+(const Coordinates& other) const {
-            return Coordinates(y + other.y, x + other.x);
-        }
+        bool operator==(const Coordinates&) const;
+        bool operator!=(const Coordinates&) const;
+        Coordinates operator+(const Coordinates&) const;
     }; // field[y][x] - y is the row, x is the column
 };
