+++ include/sista/cursor.hpp
@@ -9,13 +9,8 @@
 
 
 namespace sista {
-    void clearScreen(bool spaces=true) {
-        if (spaces) {
-            std::cout << CLS; // Clear screen
-            std::cout << SSB; // Clear scrollback buffer
-        }
-        std::cout << TL; // Move cursor to top-left corner
-    }
+    void clearScreen();
+    void clearScreen(bool);
 
     enum EraseScreen {
         FROM_CURSOR_TO_END = 0,
@@ -50,42 +45,17 @@ namespace sista {
         unsigned short int x;
         unsigned short int y;
 
-        Cursor(): x(0), y(0) {
-            std::cout << HIDE_CURSOR;
-            clearScreen();
-        }
-        ~Cursor() {
-            ANSI::reset();
-            std::cout << SHOW_CURSOR;
-            clearScreen();
-        }
+        Cursor();
+        ~Cursor();
 
-        void set(unsigned short int y_, unsigned short int x_) {
-            std::cout << CSI << y_ << ";" << x_ << CHA;
-        }
-        void set(sista::Coordinates coordinates_) {
-            this->set(coordinates_.y + 3, coordinates_.x + 2);
-        }
+        void set(unsigned short int, unsigned short int);
+        void set(sista::Coordinates);
 
-        void eraseScreen(EraseScreen eraseScreen_) {
-            std::cout << CSI << eraseScreen_ << "J";
-        }
-        void eraseLine(EraseLine eraseLine_, bool moveCursor=true) {
-            std::cout << CSI << eraseLine_ << "K";
-            if (moveCursor) {
-                this->set(this->y, 0);
-                std::cout << '\r';
-            }
-        }
+        void eraseScreen(EraseScreen);
+        void eraseLine(EraseLine, bool);
 
-        void move(MoveCursor moveCursor_, unsigned short int n=1) {
-            std::cout << CSI << n << (char)moveCursor_;
-        }
-        void move(MoveCursorDEC moveCursorDEC_) {
-            std::cout << ESC << ' ' << moveCursorDEC_;
-        }
-        void move(MoveCursorSCO moveCursorSCO_) {
-            std::cout << ESC << ' ' << moveCursorSCO_;
-        }
+        void move(MoveCursor, unsigned short int);
+        void move(MoveCursorDEC);
+        void move(MoveCursorSCO);
     };
 };
